refacto + javadoc + cleaning
