+++ backend/src/main/java/fr/projet/diginamic/backend/controllers/MissionController.java
@@ -1,4 +1,4 @@
-package fr.projet.diginamic.backend.controllers.admins;
+package fr.projet.diginamic.backend.controllers;
 
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -16,17 +16,21 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import fr.projet.diginamic.backend.dtos.BountyReportDTO;
+import fr.projet.diginamic.backend.dtos.CreateMissionDTO;
 import fr.projet.diginamic.backend.dtos.DisplayedMissionDTO;
 import fr.projet.diginamic.backend.entities.Mission;
 import fr.projet.diginamic.backend.services.CSVGenerationService;
@@ -42,12 +46,14 @@ import jakarta.validation.Valid;
 
 
 /**
- * Controller class for handling requests for admins' missions endpoint /api/admins/missions. This class handles CRUD operations for missions within the system, permitting the
+ * Controller class for handling requests for the /missions endpoint. This class
+ * handles CRUD operations for missions within the system, permitting the
  * creation, retrieval, update, and deletion of mission entries.
  */
+
 @RestController
-@RequestMapping("/api/admins/missions")
-public class MissionControllerAdmin {
+@RequestMapping("/api/missions")
+public class MissionController {
 
 	@Autowired
 	private MissionService missionService;
@@ -55,10 +61,119 @@ public class MissionControllerAdmin {
 	@Autowired
     private CSVGenerationService csvGenerationService;
 
+// ---------------------------------- CREATE MISSIONS ----------------------------------
 
-	// -------------------- GET ALL MISSIONS FOR CONNECTED ADMIN --------------------
-	
 	/**
+	 * Create a new mission with detailed validation error response.
+	 *
+	 * @param mission The mission details to create.
+	 * @param result  The binding result that holds the validation errors.
+	 * @return The created mission or validation errors.
+	 */
+	@Operation(
+        summary = "Create a new mission",
+        description = "Create a new mission with detailed validation error response. The mission details must be provided in the request body."
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "201", description = "Mission created successfully",
+            content = { @Content(mediaType = "application/json",
+            schema = @Schema(implementation = CreateMissionDTO.class)) }),
+        @ApiResponse(responseCode = "400", description = "Invalid input",
+            content = @Content),
+        @ApiResponse(responseCode = "500", description = "Internal server error",
+            content = @Content)
+    })
+	@PostMapping
+	public ResponseEntity<?> createMission(@Valid @RequestBody CreateMissionDTO mission, BindingResult result) {
+		if (result.hasErrors()) {
+			Map<String, String> errors = new HashMap<>();
+			result.getAllErrors().forEach(error -> {
+				String fieldName = ((FieldError) error).getField();
+				String errorMessage = error.getDefaultMessage();
+				errors.put(fieldName, errorMessage);
+			});
+			return ResponseEntity.badRequest().body(errors);
+		}
+	
+		try {
+			DisplayedMissionDTO savedMission = missionService.createMission(mission);
+			return ResponseEntity.status(HttpStatus.CREATED).body(savedMission);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred.");
+        }
+	}
+
+// ------------------------ GET ALL MISSIONS FOR CONNECTED USER ------------------------
+
+	/**
+	 * Retrieves a paginated list of missions based on various filtering and sorting
+	 * criteria.
+	 * This endpoint supports pagination, sorting, and dynamic filtering to provide
+	 * a flexible retrieval
+	 * of mission data. The method allows filtering by mission status, nature, and a
+	 * search term that can
+	 * match either the username of the assigned user or the mission label.
+	 *
+	 * @param page            The page number to retrieve, with a default value of 0
+	 *                        if not specified.
+	 * @param size            The size of the page to retrieve, with a default value
+	 *                        of 10 if not specified.
+	 * @param order           The direction of sorting (asc or desc), with 'asc' as
+	 *                        the default value.
+	 * @param sortField       The field on which to sort the results, with
+	 *                        'startDate' as the default field.
+	 * @param status          Optional filter to limit results to missions with a
+	 *                        specific status.
+	 * @param natureMission   Optional filter to limit results to missions of a
+	 *                        specific nature.
+	 * @param userNameOrLabel Optional search term that matches either the username
+	 *                        of the mission assignee or the mission label.
+	 * @return A {@link ResponseEntity} object containing a {@link Page} of
+	 *         {@link Mission} objects
+	 *         that match the specified criteria. The response is always OK (200),
+	 *         even if no missions match the filters.
+	 */
+
+	 @Operation(
+        summary = "Get a paginated list of all missions of the connected user",
+        description = "Retrieve a paginated list of missions based on various filtering and sorting criteria. Supports pagination, sorting, and dynamic filtering."
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "List of missions",
+            content = { @Content(mediaType = "application/json",
+            schema = @Schema(implementation = Page.class)) }),
+        @ApiResponse(responseCode = "400", description = "Invalid input",
+            content = @Content),
+        @ApiResponse(responseCode = "500", description = "Internal server error",
+            content = @Content)
+    })
+	@GetMapping
+	public ResponseEntity<?> getAllMissionsWithSpecsForConnectedUser(
+			@RequestParam(value = "page", defaultValue = "0") int page,
+			@RequestParam(value = "size", defaultValue = "10") int size,
+			@RequestParam(value = "order", defaultValue = "asc") String order,
+			@RequestParam(value = "sort", defaultValue = "startDate") String sortField,
+			@RequestParam(value = "status", required = false) String status,
+			@RequestParam(value = "nature", required = false) String natureMission,
+			@RequestParam(value = "searchbar", required = false) String userNameOrLabel) {
+
+        try {
+			Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.fromString(order), sortField));
+			Page<DisplayedMissionDTO> missions = missionService.findAllMissionsWithSpecsForCurrentUser(status, natureMission,
+					userNameOrLabel,
+					pageable);
+            return ResponseEntity.ok(missions);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred: " + e);
+        }
+	}
+
+	// -------------------- GET ALL MISSIONS OF MANAGERS COLLABORATORS  --------------------
+	 /**
 	 * Retrieves a paginated list of missions based on various filtering and sorting
 	 * criteria.
 	 * This endpoint supports pagination, sorting, and dynamic filtering to provide
@@ -100,8 +215,9 @@ public class MissionControllerAdmin {
         @ApiResponse(responseCode = "500", description = "Internal server error",
             content = @Content)
     })
-	@GetMapping
-	public ResponseEntity<?> getAllMissionsWithSpecsForAdmin(
+	@GetMapping("/collaborators")
+	public ResponseEntity<?> getAllMissionsWithSpecsForManagers(
+            @PathVariable Long id,
 			@RequestParam(value = "page", defaultValue = "0") int page,
 			@RequestParam(value = "size", defaultValue = "10") int size,
 			@RequestParam(value = "order", defaultValue = "asc") String order,
@@ -112,17 +228,16 @@ public class MissionControllerAdmin {
 
         try {
 			Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.fromString(order), sortField));
-			Page<DisplayedMissionDTO> missions = missionService.findAllMissionsWithSpecsForAdmin(status, natureMission,
-					userNameOrLabel,
-					pageable);
+			Page<DisplayedMissionDTO> missions = missionService.findAllMissionsWithSpecsByManagerId(id, status, natureMission, userNameOrLabel, pageable);
             return ResponseEntity.ok(missions);
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred.");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred." + e);
         }
 	}
 
+// -------------------------------- GET ONE MISSION BY ID  --------------------------------
 	/**
 	 * Retrieve a single mission by its ID.
 	 * 
@@ -147,10 +262,11 @@ public class MissionControllerAdmin {
         } catch (EntityNotFoundException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred.");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred : " + e);
         }
 	}
 
+	// --------------------------- UPDATE ONE MISSION BY ID  ---------------------------
 	/**
 	 * Update an existing mission.
 	 * 
@@ -192,12 +308,10 @@ public class MissionControllerAdmin {
         } catch (IllegalArgumentException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred.");
-        }
-
-		
-	}
-
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred: " + e);
+        }		
+	}	
+// ----------------------------- UPDATE ONE MISSION'STATUS -------------------------------
 	/**
 	 * Updates the status of a mission. This endpoint is restricted to managers and
 	 * administrators.
@@ -235,9 +349,11 @@ public class MissionControllerAdmin {
 			return ResponseEntity.badRequest().body(e.getMessage());
 		} catch (EntityNotFoundException e) {
 			return ResponseEntity.notFound().build();
-		}
+		}catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred: " + e);
+        }
 	}
-
+	// ----------------------------- DELETE ONE MISSION -------------------------------
 	/**
 	 * Delete a mission by its ID.
 	 * 
@@ -266,7 +382,17 @@ public class MissionControllerAdmin {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
 	}
-
+	// --------------------------- EXPORT BOUNTIES TO CSV -----------------------------
+	
+	 /**
+     * Exports the mission bounties report to a CSV file.
+     * This endpoint generates a CSV file containing the bounties report for all missions.
+     * The generated CSV file includes a summary of the bounties per month.
+     *
+     * @param response The HttpServletResponse object used to set the response headers and write the CSV file.
+     * @return A ResponseEntity indicating the status of the CSV generation.
+     * @throws IOException if there is an issue writing to the output stream.
+     */
 	@Operation(summary = "Export mission bounties to CSV")
     @ApiResponses(value = {
         @ApiResponse(responseCode = "200", description = "CSV generated successfully"),
@@ -292,5 +418,39 @@ public class MissionControllerAdmin {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
     }
-}
 
+	// ----------------------------- GET BOUNTIES REPORT -------------------------------
+	/**
+     * Retrieves the mission bounties report of the year for the specified user.
+     * 
+     * @param userId The ID of the user whose mission bounties report is to be retrieved.
+     * @return A ResponseEntity containing the bounties report if found, or an appropriate error message.
+     */
+    @Operation(
+        summary = "Get mission bounties report of the year",
+        description = "Retrieve the mission bounties report for the specified user for the current year."
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "Bounties report retrieved successfully", 
+            content = @Content(mediaType = "application/json", schema = @Schema(implementation = BountyReportDTO.class))),
+        @ApiResponse(responseCode = "400", description = "Invalid input", 
+            content = @Content),
+        @ApiResponse(responseCode = "404", description = "User or bounties not found", 
+            content = @Content),
+        @ApiResponse(responseCode = "500", description = "Internal server error", 
+            content = @Content)
+    })
+	@GetMapping("/bounties")
+	public ResponseEntity<?> getMissionBountiesOfTheYear(@PathVariable Long userId){
+		try{			
+			return ResponseEntity.ok(missionService.getBountiesReportForUser(userId));
+
+		} catch (IllegalArgumentException e) {
+			return ResponseEntity.badRequest().body(e.getMessage());
+		} catch (EntityNotFoundException e) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User or bounties not found: " + e.getMessage());
+		}catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred: " + e);
+        }
+	}
+}
