+++ backend/src/main/java/fr/projet/diginamic/backend/repositories/MissionRepository.java
@@ -71,11 +71,31 @@ public interface MissionRepository extends JpaRepository<Mission, Long>, JpaSpec
 	 */
 	Page<Mission> findAll(Pageable pageable);
 
+	 /**
+     * Retrieves a paginated list of missions for a specific user, filtered by a given specification.
+     * 
+     * @param userId    The ID of the user whose missions are to be retrieved.
+     * @param spec      The specification containing the filters to be applied to the query.
+     * @param pageable  The pagination information.
+     * @return A page of missions that match the given user ID and specification.
+     */
 	Page<Mission> findByUserId(Long userId, Specification<Mission> spec, Pageable pageable);
 
+	   /**
+     * Retrieves a list of missions for a specific user.
+     * 
+     * @param userId The ID of the user whose missions are to be retrieved.
+     * @return A list of missions associated with the given user ID.
+     */
 	List<Mission> findByUserId(Long userId);
 	
+	 /**
+     * Retrieves a paginated list of missions for users managed by a specific manager.
+     * 
+     * @param managerId The ID of the manager whose subordinates' missions are to be retrieved.
+     * @param pageable  The pagination information.
+     * @return A page of missions associated with the subordinates of the given manager.
+     */
 	Page<Mission> findAllByUser_Manager_Id(Long managerId, Pageable pageable);
-
 
 }
