+++ backend/src/main/java/fr/projet/diginamic/backend/services/CSVGenerationService.java
@@ -19,16 +19,24 @@ import org.springframework.stereotype.Service;
 import fr.projet.diginamic.backend.dtos.DisplayedMissionDTO;
 import fr.projet.diginamic.backend.utils.CalculateMissionPricing;
 
-
+/**
+ * Service class responsible for generating CSV reports related to missions and their bounties.
+ * This service utilizes mission data to create detailed CSV reports, summarizing the bounties awarded
+ * per month. It supports writing the generated CSV data directly to an output stream.
+ */
 @Service
 public class CSVGenerationService {
 
     @Autowired
     CalculateMissionPricing calculateMissionPricing;
 
-    /**
+  /**
      * Generates a CSV report of bounties per month for a given list of missions.
+     * This method processes the provided list of missions, summarizes the bounty amounts per month,
+     * and writes the results to the provided output stream in CSV format. The CSV file includes headers
+     * and skips empty header records.
      * 
+     * @param titleCSV     the title for the CSV report.
      * @param missions     the list of displayed mission DTOs.
      * @param outputStream the output stream to write the CSV file to.
      * @throws IOException if there is an issue writing to the stream.
@@ -62,24 +70,4 @@ public class CSVGenerationService {
             throw new IOException("Failed to generate CSV report", e);
         }
     }
-
-    // /**
-    //  * Summarizes bounties by month from a list of mission DTOs.
-    //  * 
-    //  * @param missions List of DisplayedMissionDTO
-    //  * @return A map with month as key and sum of bounties as value.
-    //  */
-    // private Map<String, Double> summarizeBountiesPerMonth(List<DisplayedMissionDTO> missions) {
-    //     Map<String, Double> monthSum = new HashMap<>();
-    //     for (DisplayedMissionDTO mission : missions) {
-    //         String monthKey = new SimpleDateFormat("MMMM", Locale.FRENCH).format(mission.getStartDate()).toUpperCase();
-    //         if (monthSum.containsKey(monthKey)){
-    //             Double bountiesSum = monthSum.get(monthKey) + mission.getBountyAmount() + 1;
-    //             monthSum.put(monthKey, bountiesSum);       
-    //         } else {
-    //             monthSum.put(monthKey, mission.getBountyAmount());
-    //         }
-    //     }
-    //     return monthSum;
-    // }
 }
